requirements.txt
asgiref==3.2.10
Django==3.0.7
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
importlib-metadata==1.7.0
Markdown==3.2.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
zipp==3.1.0
